Google sign-in methods added, but non-functional.
Need to refactor the commented out methods into a separate fragment
and plan to isolate the Google sign in API Client from the maps API
Client. Isolating the sign in code should make managing the lifecycle
of the required sign in objects easier instead of having to always
wait for the maps API Client to spin up.