Add ApiRequestManager class and define listener interface.
Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.Add ApiRequestManager
class and define listener interface.

Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.Add ApiRequestManager
class and define listener interface.

Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.Add ApiRequestManager
class and define listener interface.

Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.Add ApiRequestManager
class and define listener interface.

Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.Add ApiRequestManager
class and define listener interface.

Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.Add ApiRequestManager
class and define listener interface.

Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.Add ApiRequestManager
class and define listener interface.

Additionally adds the proper dependencies, permissions, and
configurations to allow for using and setting up the Volley
library.

This commit adds a required key to gradle.properties. The ApiUrl
value defines the base url for the backend server.