Add refresh button and refactor marker loading code.
Add an async class that manages the querying of the API to get a
list of all creature markers. The main thread then iterates over
the generated list of MarkerOptions instances to place them on the
map. The code to load markers is placed in a seperate method so
that it can be not only run at startup, but called easily whenever
the refresh action is used.

TODO:
 - Cache fetched information in the database.
 - Only query for information we don't have cached to minimize
     network usage.